Fixes scan_re (#8)
* Fixes scan_re

the previous regex had some bugs that let it match the too many things
This changes it to match:
https://docs.opendata.aws/noaa-nexrad/readme.html

```All files in the archive use the same compressed format (.gz). The data file names are, for example, KAKQ20010101_080138.gz. The file naming convention is:

GGGGYYYYMMDD_TTTTTT

Where:

GGGG = Ground station ID (map of ground stations)
YYYY = year
MM = month
DD = day
TTTTTT = time when data started to be collected (GMT)
Note that the 2015 files have an additional field on the file name. It adds “_V06” to the end of the file name. An example is KABX20150303_001050_V06.gz.```

* better fix

optional groups for V0X and .gz

Co-authored-by: Aaron Anderson <aarande@users.noreply.github.com>